Track datastream access/views+downloads directly via the HTTP API
Because of the way Islandora shuts down page loads when downloading/viewing datastreams, we can not attach our Async JS tracker there, so we hack a bit and implement the only hook invoked during datastream viewing (meant for altering file names, but who cares right?).
This code will try to contact Matomo via the HTTP API, passing a download url (real, not aliased path) and the right PID to the second dimension of the main site. @TODO. Should we also do this like this for authors? The need here is really just for tracking downloads for a certain PID, but we could eventually move all that is JS for file downloads into this. That would sadlty exclude the metadata downloads which are not datastreams. @noahwsmith we need a similar hook (but not for datastreams of course) in the metatadata download module you built. If you want that to be counted as a download via this mechanism. If that module provides a hook i can implement it here.